+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Move identification of run mode to settings instead of container definition
 - Pass auth and session as request attribute instead of using container
+- Use gatekeeper middleware instead of wrapped check_admin() action
 
 ## [3.4.0] - 2024-04-26 Basic json API + environment vars + clean-up
 ### Added